estok-py/requirements.txt
Flask
psycopg2-binary
requests
python-dotenv
Flask-SQLAlchemy
